auth/requirements.txt
Flask==2.1.2
Flask-PyMongo==2.3.0
PyJWT==2.3.0
gunicorn==20.1.0
python-dotenv==0.19.2
Werkzeug>=2.0